Continuing work on bg image carousel
