+++ package.json
@@ -28,11 +28,12 @@
     "cross-env": "^7.0.3",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "framer-motion": "^11.1.9",
     "jose": "^5.2.3",
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.3.0",
     "mongoose": "^8.1.1",
-    "next": "^14.1.0",
+    "next": "^14.1.1",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.0.1",
